Add warnings and reformat try_new()
