+++ kani-driver/Cargo.toml
@@ -37,6 +37,7 @@ strum_macros = {version = "0.24.0"}
 tracing = {version = "0.1", features = ["max_level_trace", "release_max_level_debug"]}
 tracing-subscriber = {version = "0.3.8", features = ["env-filter", "json", "fmt"]}
 tracing-tree = "0.2.2"
+rand = "0.8"
 which = "4.4.0"
 
 # A good set of suggested dependencies can be found in rustup:
