+++ kani-driver/src/util.rs
@@ -14,6 +14,84 @@ use std::ffi::OsString;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 
+pub mod tempfile {
+    use std::{
+        env,
+        fs::{self, rename, File},
+        io::{BufWriter, Error, Write},
+        path::PathBuf,
+    };
+
+    use crate::util;
+    use ::rand;
+    use anyhow::Context;
+    use rand::Rng;
+
+    /// Handle a writable temporary file which will be deleted when the object is dropped.
+    /// To save the contents of the file, users can invoke `rename` which will move the file to
+    /// its new location and no further deletion will be performed.
+    pub struct TempFile {
+        pub file: File,
+        pub temp_path: PathBuf,
+        pub writer: Option<BufWriter<File>>,
+        renamed: bool,
+    }
+
+    impl TempFile {
+        /// Create a temp file
+        pub fn try_new(suffix_name: &str) -> Result<Self, Error> {
+            let mut temp_path = env::temp_dir();
+
+            // Generate a unique name for the temporary directory
+            let hash: u32 = rand::thread_rng().gen();
+            let file_name: &str = &format!("kani_tmp_{hash}_{suffix_name}");
+
+            temp_path.push(file_name);
+            let temp_file = File::create(&temp_path)?;
+            let writer = BufWriter::new(temp_file.try_clone()?);
+
+            Ok(Self { file: temp_file, temp_path, writer: Some(writer), renamed: false })
+        }
+
+        /// Rename the temporary file to the new path, replacing the original file if the path points to a file that already exists.
+        pub fn rename(mut self, source_path: &str) -> Result<(), String> {
+            // flush here
+            self.writer.as_mut().unwrap().flush().unwrap();
+            self.writer = None;
+            // Renames are usually automic, so we won't corrupt the user's source file during a crash.
+            rename(&self.temp_path, source_path)
+                .with_context(|| format!("Error renaming file {}", self.temp_path.display()))
+                .unwrap();
+            self.renamed = true;
+            Ok(())
+        }
+    }
+
+    /// Ensure that the bufwriter is flushed and temp variables are dropped
+    /// everytime the tempfile is out of scope
+    /// note: the fields for the struct are dropped automatically by destructor
+    impl Drop for TempFile {
+        fn drop(&mut self) {
+            // if writer is not flushed, flush it
+            if self.writer.as_ref().is_some() {
+                // couldn't use ? as drop does not handle returns
+                if let Err(e) = self.writer.as_mut().unwrap().flush() {
+                    util::warning(
+                        format!("failed to flush {}: {e}", self.temp_path.display()).as_str(),
+                    );
+                }
+                self.writer = None;
+            }
+
+            if !self.renamed {
+                if let Err(_e) = fs::remove_file(&self.temp_path.clone()) {
+                    util::warning(&format!("Error removing file {}", self.temp_path.display()));
+                }
+            }
+        }
+    }
+}
+
 /// Replace an extension with another one, in a new PathBuf. (See tests for examples)
 pub fn alter_extension(path: &Path, ext: &str) -> PathBuf {
     path.with_extension(ext)
